requirements.txt
# Install PyTorch with CUDA first:
#   pip install --pre torch --index-url https://download.pytorch.org/whl/nightly/cu128
#
# Then install these:
gradio>=4.0
heretic-llm>=1.2.0
accelerate~=1.10
bitsandbytes~=0.45
datasets~=4.0
huggingface-hub~=0.34
optuna~=4.5
peft~=0.14
pydantic-settings~=2.10
protobuf>=3.20
sentencepiece>=0.2
tiktoken>=0.7
transformers~=4.57
rich~=14.1
